[refactor] store 생성, 사가 연결 부분을 store/index.ts로 분리
