+++ client/src/index.tsx
@@ -1,21 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { configureStore } from '@reduxjs/toolkit';
-import createSagaMiddleware from 'redux-saga';
 import { Provider } from 'react-redux';
 import App from '@/App';
-import rootReducer from '@/store/modules';
-import sagas from '@/store/sagas';
-
-const sagaMiddleware = createSagaMiddleware();
-
-const store = configureStore({
-  reducer: rootReducer,
-  middleware: [sagaMiddleware],
-  devTools: process.env.MODE === 'dev',
-});
-
-sagaMiddleware.run(sagas);
+import store from '@/store';
 
 const rootElement = document.getElementById('root');
