+++ client/src/store/index.ts
@@ -1,0 +1,16 @@
+import { configureStore } from '@reduxjs/toolkit';
+import createSagaMiddleware from 'redux-saga';
+import rootReducer from '@/store/modules';
+import sagas from '@/store/sagas';
+
+const sagaMiddleware = createSagaMiddleware();
+
+const store = configureStore({
+  reducer: rootReducer,
+  middleware: [sagaMiddleware],
+  devTools: process.env.MODE === 'dev',
+});
+
+sagaMiddleware.run(sagas);
+
+export default store;
